src/faker/requirements.txt
Faker==37.5.3
psycopg2-binary==2.9.10

